Finish inefficient algorithm. Add cylinder test. Add iteration test.

+++ include/Cylinder.hpp
@@ -6,13 +6,13 @@
 namespace cadcam {
 class Cylinder : public Volume {
 public:
-    Cylinder(const mwTPoint3d<double>& start, const mwTPoint3d<double>& finish, float radius);
+    Cylinder(const mwTPoint3d<double>& start, const mwTPoint3d<double>& finish, double radius);
     bool ContainsPoint(const mwTPoint3d<double> &point) const override;
 
 private:
     mwTPoint3d<double> start_;
     mwTPoint3d<double> finish_;
-    float radius_;
+    double radius_;
 
     double inverseEquations[2][2] {};
 };
